+++ src/TRestDetectorGarfieldDriftProcess.cxx
@@ -330,7 +330,7 @@ void TRestDetectorGarfieldDriftProcess::InitProcess() {
 Int_t TRestDetectorGarfieldDriftProcess::FindModule(Int_t readoutPlane, Double_t x, Double_t y) {
     // TODO verify this
     TRestDetectorReadoutPlane* plane = fReadout->GetReadoutPlane(readoutPlane);
-    for (int md = 0; md < plane->GetNumberOfModules(); md++)
+    for (size_t md = 0; md < plane->GetNumberOfModules(); md++)
         if ((*plane)[md].isInside(x, y)) return md;
 
     return -1;
