+++ src/TRestDetectorSingleChannelAnalysisProcess.cxx
@@ -48,7 +48,7 @@ void TRestDetectorSingleChannelAnalysisProcess::InitProcess() {
     } else {
         for (int i = 0; i < fReadout->GetNumberOfReadoutPlanes(); i++) {
             auto plane = fReadout->GetReadoutPlane(i);
-            for (int j = 0; j < plane->GetNumberOfModules(); j++) {
+            for (size_t j = 0; j < plane->GetNumberOfModules(); j++) {
                 auto mod = plane->GetModule(j);
                 for (int k = 0; k < mod->GetNumberOfChannels(); k++) {
                     auto channel = mod->GetChannel(k);
